chore(web): resolve requested changes from PR #38
