+++ apps/web/tailwind.config.ts
@@ -7,6 +7,7 @@ const config = {
 		'./components/**/*.{ts,tsx}',
 		'./app/**/*.{ts,tsx}',
 		'./src/**/*.{ts,tsx}',
+		'./lib/**/*.{ts,tsx}',
 	],
 	prefix: '',
 	theme: {
